requirements.txt
pyowm
pygame
newsapi-python